BrowserDetector code in place. Curtain demo updated accordingly.
BrowserDetector has been updated and fixed to check for the current Custom Filter syntax instead for specific UA versions.
Stub files have been put in external, accessible, public directories to minimize the risk of directory lists with misconfigured
httpd.
jQuery-min is now part of the resources, although in future we might decide to use a different way to check for the feature.

More samples will be updated shortly.